Time to wait
A time to wait boolean file - will return true only when the passed value for desired time has been reached or passed.  This file is currently very weak, as it is possible for it to run infinitely.
I still need to go back and double check the Java functionality and syntax once at I’m at my desktop.